requirements.txt
requests
pyyaml
pillow
numpy